requirements.txt
# Core GUI and Interface
customtkinter
pillow

# Data Processing and Analysis
numpy
pandas
matplotlib
seaborn

# Email and Gmail API
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
beautifulsoup4

# AI and Language Processing
boto3
faiss-cpu

# Speech Recognition and TTS
SpeechRecognition
pyttsx3

# Utilities
python-dateutil
python-dotenv
tzlocal
chime

# Web Interface
streamlit

# Database
sqlalchemy
psycopg2-binary
pymongo